Use a volume scale from 0.0 to 1.0. Fixed copying of remote cover images.
2007-05-03  Paolo Bacchilega  <paobac@cvs.gnome.org>

	* many_files: Use a volume scale from 0.0 to 1.0.  Fixed copying of
	remote cover images.  Save the cover at the original size, resize
	the cover before displaying it.

svn path=/trunk/; revision=251

+++ src/goo-player.h
@@ -114,9 +114,9 @@ const char *     goo_player_get_location        (GooPlayer       *player);
 const char *     goo_player_get_title           (GooPlayer       *player);
 const char *     goo_player_get_artist          (GooPlayer       *player);
 int              goo_player_get_year            (GooPlayer       *player);
-int              goo_player_get_volume          (GooPlayer       *player);
+double           goo_player_get_volume          (GooPlayer       *player);
 void             goo_player_set_volume          (GooPlayer       *player,
-						 int              vol);
+						 double           vol);
 gboolean         goo_player_get_is_busy         (GooPlayer       *player);
 
 /**/
